8: JWTAuthentication returns None instead of raises
This modifies the JWTAuthentication provider to return None instead of raising an Exception if the token cannot be created, or is invalid.  This matches the expectations outlined in #8, and closes #8.

ConfigurationErrors are special cased and allowed to bubble up.  It seems better to have the server crash quickly then to silently not provide authentication for a static and fixable issue.